requirements.txt
customtkinter==5.2.1
tkinterdnd2==0.3.0
pdf2image
Pillow
pymupdf 